+++ ChangeLog.md
@@ -69,7 +69,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 * This component is no longer supported on PHP 8.2
 
-[8.2.0]: https://github.com/sebastianbergmann/exporter/compare/8.1.1...main
+[8.2.0]: https://github.com/sebastianbergmann/exporter/compare/8.1.1...8.2.0
 [8.1.1]: https://github.com/sebastianbergmann/exporter/compare/8.1.0...8.1.1
 [8.1.0]: https://github.com/sebastianbergmann/exporter/compare/8.0.3...8.1.0
 [8.0.3]: https://github.com/sebastianbergmann/exporter/compare/8.0.2...8.0.3